Changed attempt to mediumint(6) to allow for larger values to be stored.

+++ SQL/Migrations/attempt_value_fix.sql
@@ -1,2 +1,2 @@
-alter table pfiles_inv modify `Attempt` smallint(1) default '0';
-alter table pfiles_worn modify `Attempt` smallint(1) default '0';
+alter table pfiles_inv modify `Attempt` mediumint(6) default '0';
+alter table pfiles_worn modify `Attempt` mediumint(6) default '0';
